Move Lens Reader/State extensions to optis-mtl
