+++ modules/optics/arrow-optics/src/main/kotlin/arrow/optics/Lens.kt
@@ -8,8 +8,6 @@ import arrow.core.Some
 import arrow.core.Tuple2
 import arrow.core.identity
 import arrow.core.toT
-import arrow.mtl.Reader
-import arrow.mtl.State
 import arrow.mtl.map
 import arrow.higherkind
 import arrow.typeclasses.Applicative
@@ -234,73 +232,4 @@ interface PLens<S, T, A, B> : PLensOf<S, T, A, B> {
    * Verify if the focus of a [PLens] satisfies the predicate
    */
   fun exist(s: S, p: (A) -> Boolean): Boolean = p(get(s))
-
-  /**
-   * Extracts the value viewed through the [get] function.
-   */
-  fun ask(): Reader<S, A> = Reader(::get)
-
-  /**
-   * Transforms a [PLens] into a [Reader]. Alias for [ask].
-   */
-  fun toReader(): Reader<S, A> = ask()
-
-  /**
-   * Extracts the value viewed through the [get] and applies [f] to it.
-   *
-   * @param f function to apply to the focus.
-   */
-  fun <C> asks(f: (A) -> C): Reader<S, C> = ask().map(f)
-
-  /**
-   * Extracts the focus [A] viewed through the [PLens].
-   */
-  fun extract(): State<S, A> = State { s -> Tuple2(s, get(s)) }
-
-  /**
-   * Transforms a [PLens] into a [State].
-   * Alias for [extract].
-   */
-  fun toState(): State<S, A> = extract()
-
-  /**
-   * Extracts and maps the focus [A] viewed through the [PLens] and applies [f] to it.
-   */
-  fun <C> extractMap(f: (A) -> C): State<S, C> = extract().map(f)
 }
-
-/**
- * Update the focus [A] viewed through the [Lens] and returns its *new* value.
- */
-fun <S, A> Lens<S, A>.update(f: (A) -> A): State<S, A> = State { s ->
-  val b = f(get(s))
-  Tuple2(set(s, b), b)
-}
-
-/**
- * Update the focus [A] viewed through the [Lens] and returns its *old* value.
- */
-fun <S, A> Lens<S, A>.updateOld(f: (A) -> A): State<S, A> = State { s ->
-  Tuple2(modify(s, f), get(s))
-}
-
-/**
- * Modify the focus [A] viewed through the [Lens] and ignores both values.
- */
-fun <S, A> Lens<S, A>.update_(f: (A) -> A): State<S, Unit> =
-  State { s -> Tuple2(modify(s, f), Unit) }
-
-/**
- * Assign the focus [A] viewed through the [Lens] and returns its *new* value.
- */
-fun <S, A> Lens<S, A>.assign(a: A): State<S, A> = update { _ -> a }
-
-/**
- * Assign the value focus [A] through the [Lens] and returns its *old* value.
- */
-fun <S, A> Lens<S, A>.assignOld(a: A): State<S, A> = updateOld { _ -> a }
-
-/**
- * Assign the focus [A] viewed through the [Lens] and ignores both values.
- */
-fun <S, A> Lens<S, A>.assign_(a: A): State<S, Unit> = update_ { _ -> a }
